app/requirements.txt
cryptography==46.0.3
fastapi==0.127.0
passlib==1.7.4
pydantic==2.12.5
python-dotenv==1.2.1
python_jose==3.5.0
SQLAlchemy==2.0.45
